saga_api: gcc warning int-to-pointer-cast

+++ saga-gis/src/saga_core/saga_api/parameters.h
@@ -79,7 +79,7 @@
 #include "shapes.h"
 #include "tin.h"
 #include "pointcloud.h"
-
+#include "stdint.h"
 
 ///////////////////////////////////////////////////////////
 //														 //
@@ -789,7 +789,7 @@ public:
 
 	virtual bool				is_Valid				(void);
 
-	virtual bool				Set_Value				(int   Value)	{	return( Set_Value((void *)Value) );	}
+	virtual bool				Set_Value				(uintptr_t Value)	{	return( Set_Value((void *)Value) );	}
 	virtual bool				Set_Value				(void *Value);
 
 	virtual void *				asPointer				(void)			{	return( m_pDataObject );	}
